chore:合併vuex module 為index, 狀態判斷調整
