+++ src/pages/homepage/utils/statusUtils.js
@@ -1,7 +1,7 @@
 export function setPendingInterviewStatus(status, interviewDate) {
   const today = new Date();
   const interview = new Date(interviewDate);
-  if (status === "無") {
+  if (status === "none" || status === "無") {
     if (interview < today) {
       return "等待結果"; // 面試日期早於今天
     } else {
